CDAP-15266 Changed Sink config properties to match current plugin standards
