+++ docs/MongoDB-batchsink.md
@@ -8,7 +8,22 @@ Converts a StructuredRecord into a BSONWritable and then writes it to a MongoDB 
 
 Configuration
 -------------
-**referenceName:** This will be used to uniquely identify this sink for lineage, annotating metadata, etc.
+**Reference Name:** Name used to uniquely identify this source for lineage, annotating metadata, etc.
 
-**connectionString:** MongoDB Connection String. Example: `mongodb://localhost:27017/analytics.users` (Macro-enabled)
-[Reference](http://docs.mongodb.org/manual/reference/connection-string)
+**Host:** Host that MongoDB is running on.
+
+**Port:** Port that MongoDB is listening to.
+
+**Database:** MongoDB database name.
+
+**Collection:** Name of the database collection to write to.
+
+**Username:** User identity for connecting to the specified database.
+
+**Password:** Password to use to connect to the specified database.
+
+**Connection Arguments:** A list of arbitrary string key/value pairs as connection arguments. See
+[Connection String Options] for a full description of these arguments.
+
+[Connection String Options]:
+https://docs.mongodb.com/manual/reference/connection-string/#connections-connection-options
